// src/api/models/update-my-info-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateMyInfoReq
 */
export interface UpdateMyInfoReq {
    /**
     * 
     * @type {string}
     * @memberof UpdateMyInfoReq
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMyInfoReq
     */
    'phoneNumber'?: string;
}